EDIT utilizer_test
