+++ utilizer_test.sh
@@ -4,4 +4,4 @@ export API_KEY=kvtkOoyqcuTgNrBqRGIhhLe766CLYbpo
 export START_PAIR=$1
 echo "API_KEY"=$API_KEY
 echo "START_PAIR="$1
-screen python test/utilizer_integration_test.py
+screen python test/utilizer_test.py
